+++ webapp/components/display-logo.js
@@ -100,33 +100,6 @@ class DisplayLogo extends ProductInfo(PolymerElement) {
     return small ? 'small' : '';
   }
 
-  displayLogo(name, labels) {
-    if (!name) {
-      return;
-    }
-    // TODO: Remove the special case for Servo when it has per-channel logos.
-    if (name !== 'servo' && labels) {
-      labels = new Set(labels);
-      let channel;
-      const candidates = ['beta', 'dev', 'canary', 'nightly', 'preview'];
-      for (const label of candidates) {
-        if (labels.has(label)) {
-          channel = label;
-          break;
-        }
-      }
-      // Fall back to treating 'experimental' as 'dev'.
-      // TODO: Remove after https://github.com/web-platform-tests/wpt.fyi/issues/1539.
-      if (!channel && labels.has('experimental')) {
-        channel = 'dev';
-      }
-      if (channel) {
-        name = `${name}-${channel}`;
-      }
-    }
-    return `/static/${name}_64x64.png`;
-  }
-
   computeSource(product, showSource) {
     if (!showSource || !product.labels) {
       return '';
